services/backtest/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backtrader==1.9.78.123
bcrypt==3.2.0
blinker==1.4
certifi==2025.8.3
chardet==4.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.4
comm==0.2.3
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
dill==0.4.0
dnspython==2.7.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fonttools==4.59.1
frozenlist==1.7.0
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
incremental==21.3.0
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.5.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyterlab_widgets==3.0.15
keyring==23.5.0
kiwisolver==1.4.9
lia-web==0.2.3
llvmlite==0.39.1
MarkupSafe==2.0.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
more-itertools==8.10.0
motor==3.7.1
multidict==6.6.4
mypy_extensions==1.1.0
netifaces==0.11.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.0
packaging==25.0
pandas==2.3.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==5.24.1
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.14.1
pyOpenSSL==21.0.0
pyparsing==3.2.3
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
ruff==0.12.11
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.15.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
strawberry-graphql==0.278.1
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
Twisted==22.1.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vectorbt==0.28.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
yarl==1.20.1
zipp==1.0.0
zope.interface==5.4.0